test(connectors): share the Glooko sync harness across service tests
The fake session, endpoint handler and publish-recording service were nested
inside the publish-failure tests, where a second sync-level test could not
reach them. The handler now serves a per-window record count so a test can
drive more than one chunk.
